+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Remove chat entry component ([PR #4512](https://github.com/alphagov/govuk_publishing_components/pull/4512))
 * Use component wrapper on skip link component ([PR #4520](https://github.com/alphagov/govuk_publishing_components/pull/4520))
 * Use component wrapper on step nav header component ([PR #4521](https://github.com/alphagov/govuk_publishing_components/pull/4521))
+* Use component wrapper on step nav related component ([PR #4522](https://github.com/alphagov/govuk_publishing_components/pull/4522))
 
 ## 46.4.0